@@ -5,24 +5,23 @@
   "private": true,
   "dependencies": {
     "@craco/craco": "^6.4.3",
-    "@types/react": "^17.0.40",
-    "body-parser": "^1.19.2",
+    "@types/react": "^18.0.12",
+    "body-parser": "^1.20.0",
     "buffer": "^6.0.3",
-    "concurrently": "^7.0.0",
+    "concurrently": "^7.2.1",
     "cross-env": "^7.0.3",
-    "dotenv": "^16.0.0",
+    "dotenv": "^16.0.1",
     "es6-promise": "^4.2.8",
-    "express": "^4.17.3",
-    "express-rate-limit": "^6.3.0",
+    "express": "^4.18.1",
+    "express-rate-limit": "^6.4.0",
     "express-secure-only": "^0.2.1",
-    "helmet": "^5.0.2",
-    "husky": "^7.0.4",
-    "ibm-watson": "^6.2.2",
+    "helmet": "^5.1.0",
+    "husky": "^8.0.1",
+    "ibm-watson": "^7.0.0",
     "isomorphic-fetch": "^3.0.0",
-    "lint-staged": "^12.3.7",
+    "lint-staged": "^13.0.0",
     "morgan": "^1.10.0",
     "process": "^0.11.10",
-    "sass": "^1.49.9",
     "stream-browserify": "^3.0.0",
     "vcap_services": "^0.7.1",
     "watson-speech": "^0.39.1"
@@ -62,25 +61,23 @@
     ]
   },
   "devDependencies": {
-    "@carbon/themes": "^10.53.1",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
-    "carbon-components": "^10.55.1",
-    "carbon-components-react": "^7.55.1",
-    "carbon-icons": "^7.0.7",
+    "@carbon/react": "^1.4.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.0",
     "carbon-themes": "^1.7.0",
-    "jest": "27.5.1",
+    "jest": "28.1.0",
     "jest-puppeteer": "^6.1.0",
-    "nodemon": "^2.0.15",
-    "prettier": "^2.5.1",
-    "puppeteer": "^13",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-json-tree": "^0.16.1",
+    "nodemon": "^2.0.16",
+    "prettier": "^2.6.2",
+    "puppeteer": "^14",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-json-tree": "^0.17.0",
     "react-json-view": "^1.21.3",
-    "react-scripts": "^5.0.0",
-    "webpack": "^5.70.0"
+    "react-scripts": "^5.0.1",
+    "sass": "^1.52.2",
+    "webpack": "^5.73.0"
   },
   "prettier": {
     "trailingComma": "all",
