@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Tile } from 'carbon-components-react';
+import { Tile } from '@carbon/react';
 
 export const Header = ({ description, links, title }) => (
   <Tile className="header">
